refactor: share cloud connector and card backup logic
The connector drawing was copied almost verbatim between the desktop and
big picture cloud saves sections, and each one repeated the same start,
upload and finish dance for a card backup. Move the connector into a
useCloudConnector hook and expose a backupCard helper from
useEmulationBackupProgress so both sections call into one place.
