+++ src/renderer/src/pages/settings/emulation/cloud-saves-section.tsx
@@ -1,10 +1,4 @@
-import {
-  useCallback,
-  useEffect,
-  useLayoutEffect,
-  useRef,
-  useState,
-} from "react";
+import { useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import {
   ClockIcon,
@@ -20,6 +14,7 @@ import {
 import { Button, ConfirmationModal } from "@renderer/components";
 import { DropdownMenu } from "@renderer/components/dropdown-menu/dropdown-menu";
 import { useToast, useUserDetails } from "@renderer/hooks";
+import { useCloudConnector } from "@renderer/hooks/use-cloud-connector";
 import { useSubscription } from "@renderer/hooks/use-subscription";
 import type {
   EmulationCloudSave,
@@ -38,16 +33,6 @@ interface Props {
   refreshKey: number;
 }
 
-const SLOT_X_RATIO = 59.5 / 411;
-const SLOT_Y_RATIO = 129 / 221;
-const BRANCH_GAP = 40;
-
-interface Connector {
-  width: number;
-  height: number;
-  path: string;
-}
-
 export function CloudSavesSection({ config, refreshKey }: Readonly<Props>) {
   const { t } = useTranslation("settings");
   const { t: tHydraCloud } = useTranslation("hydra_cloud");
@@ -62,14 +47,7 @@ export function CloudSavesSection({ config, refreshKey }: Readonly<Props>) {
   const [renameFor, setRenameFor] = useState<EmulationCloudSave | null>(null);
   const [deleteFor, setDeleteFor] = useState<EmulationCloudSave | null>(null);
 
-  const stageRef = useRef<HTMLDivElement>(null);
-  const consoleRef = useRef<HTMLDivElement>(null);
-  const gridRef = useRef<HTMLDivElement>(null);
-  const [connector, setConnector] = useState<Connector>({
-    width: 0,
-    height: 0,
-    path: "",
-  });
+  const { stageRef, consoleRef, gridRef, connector } = useCloudConnector(saves);
 
   const load = useCallback(async () => {
     if (!hasActiveSubscription) {
@@ -95,96 +73,6 @@ export function CloudSavesSection({ config, refreshKey }: Readonly<Props>) {
     showSuccessToast(t("cloud_delete_success"));
     load();
   }, [deleteFor, showSuccessToast, t, load]);
-
-  const drawConnector = useCallback(() => {
-    const stage = stageRef.current;
-    const consoleEl = consoleRef.current;
-    const grid = gridRef.current;
-    if (!stage || !consoleEl || !grid) return;
-
-    const stageRect = stage.getBoundingClientRect();
-    const consoleRect = consoleEl.getBoundingClientRect();
-    const cards = Array.from(
-      grid.querySelectorAll<HTMLElement>(".emulator-detail__cloud-card")
-    );
-    if (cards.length === 0) {
-      setConnector({
-        width: stageRect.width,
-        height: stageRect.height,
-        path: "",
-      });
-      return;
-    }
-
-    const slotX =
-      consoleRect.left - stageRect.left + SLOT_X_RATIO * consoleRect.width;
-    const slotY =
-      consoleRect.top - stageRect.top + SLOT_Y_RATIO * consoleRect.height;
-
-    const cardGeoms = cards.map((card) => {
-      const rect = card.getBoundingClientRect();
-      return {
-        x: rect.left - stageRect.left + rect.width / 2,
-        top: rect.top - stageRect.top,
-        bottom: rect.bottom - stageRect.top,
-      };
-    });
-
-    type CardGeom = (typeof cardGeoms)[number];
-
-    const rowsByTop = new Map<number, CardGeom[]>();
-    for (const geom of cardGeoms) {
-      const key = Math.round(geom.top);
-      const row = rowsByTop.get(key);
-      if (row) row.push(geom);
-      else rowsByTop.set(key, [geom]);
-    }
-    const firstRow = Array.from(rowsByTop.entries()).sort(
-      (a, b) => a[0] - b[0]
-    )[0][1];
-
-    const segments: string[] = [];
-
-    const rowTop = Math.min(...firstRow.map((c) => c.top));
-    const busY = rowTop - BRANCH_GAP;
-    const xs = firstRow.map((c) => c.x).sort((a, b) => a - b);
-    const left = Math.min(slotX, xs[0]);
-    const right = Math.max(slotX, xs[xs.length - 1]);
-    segments.push(`M ${slotX} ${slotY} L ${slotX} ${busY}`);
-    segments.push(`M ${left} ${busY} L ${right} ${busY}`);
-    for (const geom of firstRow) {
-      segments.push(`M ${geom.x} ${busY} L ${geom.x} ${geom.top}`);
-    }
-
-    const columns = new Map<number, CardGeom[]>();
-    for (const geom of cardGeoms) {
-      const key = Math.round(geom.x);
-      const column = columns.get(key);
-      if (column) column.push(geom);
-      else columns.set(key, [geom]);
-    }
-    for (const column of columns.values()) {
-      column.sort((a, b) => a.top - b.top);
-      for (let i = 1; i < column.length; i += 1) {
-        const upper = column[i - 1];
-        const lower = column[i];
-        segments.push(`M ${lower.x} ${upper.bottom} L ${lower.x} ${lower.top}`);
-      }
-    }
-
-    setConnector({
-      width: stageRect.width,
-      height: stageRect.height,
-      path: segments.join(" "),
-    });
-  }, []);
-
-  useLayoutEffect(() => {
-    drawConnector();
-    const observer = new ResizeObserver(drawConnector);
-    if (stageRef.current) observer.observe(stageRef.current);
-    return () => observer.disconnect();
-  }, [drawConnector, saves]);
 
   if (!hasActiveSubscription) {
     return (
